feat: move components in their domain folder
