+++ src/lib/components/Comic/ComicStrip.svelte
