+++ src/lib/components/Painting/PaintingFeedback.svelte
