+++ src/lib/components/Painting/PaintingIntroduction.svelte
