+++ src/lib/components/Result/BottomActions.svelte
@@ -1,5 +1,5 @@
 <script>
-    import Button from "$lib/components/Button.svelte";
+    import Button from "$lib/components/Button/Button.svelte";
     import {goto} from "$app/navigation";
 
     function goBackToHome() {
