+++ src/lib/components/Suspect/SuspectEmotion.svelte
