+++ src/routes/game/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import {selectedSuspects} from "$lib/stores/selectedSuspects";
     import {paintings} from "$lib/stores/paintings";
-    import PaintingLevel from "$lib/components/PaintingLevel.svelte";
+    import PaintingLevel from "$lib/components/Painting/PaintingLevel.svelte";
     import {goto} from "$app/navigation";
 
     let painting = 0;
